requirement.txt
asn1crypto==0.24.0
awesome-slugify==1.6.5
cffi==1.11.5
cryptography==2.3.1
Django==2.1.2
django-redis==4.10.0
idna==2.7
mysqlclient==1.3.14
pkg-resources==0.0.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.5
redis==3.1.0
regex==2018.11.22
six==1.11.0
Unidecode==0.4.21
